added more files; rewrote VariantUtils::hashCode

+++ include/lucene++/BooleanScorer.h
@@ -96,7 +96,7 @@ public:
     int32_t freq;
     LUCENE_CLASS(BucketScorer);
     float termFreq(){
-        return freq;
+        return (float)freq;
     }
